refactor(compiler): split schema and data projection ownership
