+++ packages/compiler/src/data/index.ts
@@ -1,0 +1,2 @@
+export { renderCoreDataContract } from "./generated-contract";
+export { projectDataRelations } from "./relations";
